requirements.txt
discord.py==2.0.1
pillow==9.0.1
numpy==1.23.0
